+++ src/components/Credits/Credits.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+function Credits() {
+  return (
+    <div>
+      <small>
+        Made by <a href="https://lorisbettazza.com">Loris Bettazza</a>
+        <br />
+        View{' '}
+        <a href="https://github.com/Pustur/whatsapp-chat-parser-website">
+          Source code
+        </a>
+      </small>
+    </div>
+  );
+}
+
+export default Credits;