+++ src/style.js
@@ -1,10 +1,8 @@
 import styled, { css, createGlobalStyle } from 'styled-components';
 
-import { hideText, normalizeButton, normalizeInput } from './utils/styles';
+import { normalizeInput } from './utils/styles';
 import { whatsappThemeColor } from './utils/colors';
-import { zIndex } from './utils/z-index';
 
-const buttonSize = '44px';
 const selectArrowWidth = '10px';
 const selectArrowHeight = '5px';
 const selectPadding = '0.3rem';
@@ -24,187 +22,10 @@ const inputStyles = css`
   }
 `;
 
-const GlobalStyles = createGlobalStyle`
-  *,
-  *::before,
-  *::after {
-    box-sizing: inherit;
-  }
-
-  html {
-    font-family: sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 
-      'Segoe UI Symbol', 'Noto Color Emoji';
-    box-sizing: border-box;
-  }
-
-  body {
-    margin: 0;
-    color: #333;
-  }
-
-  a {
-    text-decoration: none;
-    color: ${whatsappThemeColor};
-  }
-
-  img,
-  video,
-  audio {
-    max-width: 100%;
-  }
-
-  button {
-    cursor: pointer;
-  }
-
-  html,
-  body,
-  #root {
-    height: 100%;
-  }
-
-  @media (prefers-color-scheme: dark) {
-    body {
-      background-color: #262d31;
-      color: #ccc;
-    }
-  }
-`;
-
 const Container = styled.div`
   display: flex;
   flex-direction: column;
   min-height: 100%;
-`;
-
-const MenuOpenButton = styled.button`
-  ${normalizeButton}
-  ${hideText}
-
-  position: fixed;
-  width: ${buttonSize};
-  height: ${buttonSize};
-  left: 1rem;
-  bottom: 1rem;
-  border-radius: 50%;
-  background-color: ${whatsappThemeColor};
-
-  &::after {
-    content: '';
-    display: block;
-    position: absolute;
-    width: 16px;
-    height: 2px;
-    top: 50%;
-    left: 50%;
-    transform: translate3d(-50%, -50%, 0);
-    background-color: white;
-    box-shadow: 0 -5px 0 white, 0 5px 0 white;
-  }
-
-  @media (min-width: 700px) {
-    left: 2rem;
-    bottom: 2rem;
-  }
-`;
-
-const MenuCloseButton = styled.button`
-  ${normalizeButton}
-  ${hideText}
-
-  position: absolute;
-  width: ${buttonSize};
-  height: ${buttonSize};
-  top: 0;
-  right: 0;
-  background-color: transparent;
-  opacity: 0.5;
-  transition: opacity 0.3s ease;
-
-  &:hover,
-  &:focus {
-    opacity: 1;
-  }
-
-  &::before,
-  &::after {
-    content: '';
-    display: block;
-    position: absolute;
-    width: 20px;
-    height: 2px;
-    top: 50%;
-    left: 50%;
-    transform-origin: 50% 50%;
-    background-color: black;
-  }
-
-  &::before {
-    transform: translate3d(-50%, -50%, 0) rotate(45deg);
-  }
-
-  &::after {
-    transform: translate3d(-50%, -50%, 0) rotate(135deg);
-  }
-
-  @media (prefers-color-scheme: dark) {
-    &::before,
-    &::after {
-      background-color: white;
-    }
-  }
-`;
-
-const Overlay = styled.button`
-  ${normalizeButton}
-
-  display: block;
-  position: fixed;
-  width: 100%;
-  top: 0;
-  bottom: 0;
-  background-color: black;
-  opacity: ${props => (props.isActive ? 0.2 : 0)};
-  transition: opacity 0.3s ease;
-  z-index: ${zIndex.overlay};
-  ${props =>
-    !props.isActive &&
-    css`
-      pointer-events: none;
-    `}
-`;
-
-const Sidebar = styled.aside`
-  position: fixed;
-  width: 280px;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  background-color: white;
-  transform: translate3d(${props => (props.isOpen ? 0 : '-100%')}, 0, 0);
-  transition: transform 0.3s ease;
-  z-index: ${zIndex.sidebar};
-
-  @media (prefers-color-scheme: dark) {
-    background-color: #262d31;
-  }
-`;
-
-const SidebarContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  position: absolute;
-  top: ${buttonSize};
-  left: 0;
-  bottom: 0;
-  right: 0;
-  padding: 1rem;
-  border-top: 1px solid #eee;
-
-  @media (prefers-color-scheme: dark) {
-    border-color: #444;
-  }
 `;
 
 const Form = styled.form`
@@ -300,14 +121,62 @@ const Header = styled.header`
   }
 `;
 
+const GlobalStyles = createGlobalStyle`
+  *,
+  *::before,
+  *::after {
+    box-sizing: inherit;
+  }
+
+  html {
+    font-family: sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 
+      'Segoe UI Symbol', 'Noto Color Emoji';
+    box-sizing: border-box;
+  }
+
+  body {
+    margin: 0;
+    color: #333;
+  }
+
+  a {
+    text-decoration: none;
+    color: ${whatsappThemeColor};
+  }
+
+  img,
+  video,
+  audio {
+    max-width: 100%;
+  }
+
+  button {
+    cursor: pointer;
+  }
+
+  html,
+  body,
+  #root {
+    height: 100%;
+  }
+
+  @media (prefers-color-scheme: dark) {
+    body {
+      background-color: #262d31;
+      color: #ccc;
+    }
+  }
+
+  @media print {
+    video, audio, ${Header}, .menu-open-button {
+      display: none !important;
+    }
+  }
+`;
+
 export {
   GlobalStyles,
   Container,
-  MenuOpenButton,
-  MenuCloseButton,
-  Overlay,
-  Sidebar,
-  SidebarContainer,
   Form,
   Field,
   Label,